// apps/web/src/lib/finance/single-income-types.ts
export type SingleIncomeIssueRow = {
  issueType: string
  errorMessage: string
  projectId: string
  projectName: string
  tenantId: string
  tenantName: string
  platformTenantId: string
  customerId: string
  customerName: string
  billId: string
}

export type SingleIncomeProjectDetail = {
  projectId: string
  projectName: string
  tenantId: string
  tenantName: string
  platformTenantId: string
  customerId: string
  customerName: string
}